index: Add Index.NumNames to iterate over all names.
Index.Name() returns the name of a file, but there is no way to tell
what are valid bounds for this function. Index.Paths() returns a copy
of all names, which is expensive for big index files. NumNames()
permits cheap iteration over all names in the index, without needing
to allocate all the names up front.

README: Document this fork.